Speed improvements - fit inside a cache line
If we use a native tm struct we overfill the X64 cache line. (The issue
is that we also need the two longs to store the native timestamp).

We don't actually need the TM struct until we break the time into
components, and the relative cost of the extra copy is acceptable.

A better solution would be to break the class into two pieces (one for
the timestamp, one for the compoentisation), but that leaves with us
with a backwardsly incompatible API - so has been left until the need
arrises.